refactor(ReservationDto) : 요청 및 반환 수정
