+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,11 +1,10 @@
-package roomescape.reservation.dto.request;
+package roomescape.reservation.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
-import java.time.LocalTime;
 
 public record ReservationRequest(
         String name,
         @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
-        @JsonFormat(pattern = "HH:mm") LocalTime time) {
+        Long timeId) {
 }
