+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,0 +1,13 @@
+package roomescape.reservation.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDate;
+import roomescape.time.dto.TimeResponse;
+
+public record ReservationResponse(
+        Long id,
+        String name,
+        @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
+        TimeResponse time) {
+
+}
