--- src/main/java/roomescape/reservation/dto/response/ReservationAllResponse.java
@@ -1,13 +1,0 @@
-package roomescape.reservation.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record ReservationAllResponse(
-        Long id,
-        String name,
-        LocalDate date,
-        @JsonFormat(pattern = "HH:mm") LocalTime time) {
-
-}
